Scoring change - Testing
I am changing how scoring works. Instead of copying the score from the
literature on wikipedia, I am calling an audible and making the score
the percentage difference between the words.  If the Lev distance is 1
in a ten letter long word, you will have a 90% match.

Also added unit testing for Scoring as well as Levenshtein distance.