Removed unnecessary dependancies
Removed the timing code. That really belongs in a benchmark.
Removed the gorilla library. I wasn't using it enough to force users to
import it.
Added the _ "expvar" import to register http handlers so the library
can clip into the user's already functioning http handler. This is a
really important update. The cleo handler is delcared in the init
function.
I also renamed a few internal functions.

I created a log.fatal statement on the input file if there is any error
reading the input file since you will not be able to use the library
otherwise.